backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Affiliate Comparison Platform Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "psycopg2-binary",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "langgraph",
    "langchain",
    "langchain-core",
    "langchain-google-genai",
    "langchain-openai",
    "httpx",
    "beautifulsoup4",
    "alembic",
    "playwright",
    "supabase",
    "gotrue",
    "ddgs",
    "python-slugify",
    "python-multipart",
    "passlib[bcrypt]",
    "bcrypt",
    "boto3",
    "aiohttp",
    "apscheduler",
]


[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
